Adding in XML
SimulateMatch is now creating an XML file to save the data so it's not gone on startup. It can create the base for the XML and load the file up on startup. It may even be able to add new information into the XML file, don't know yet, testing that now

RankModel does nothing, may remove one day but leaving it in for now